Fix sign in
...

+++ KMMSEP/Scripts/app/Controllers/IndexController.js
@@ -97,16 +97,24 @@
             targetEvent: ev,
             clickOutsideToClose: true,
             fullscreen: false,
-        })
-        .then(function () {
-            $scope.userName = CommonFunctionService.getCookie('kmmsep_user').name;
+            onRemoving: function () {
+                $scope.userName = CommonFunctionService.getCookie('kmmsep_user').name;
+            }
         })
     };
 
-    $scope.signOut = function () {
-        CommonFunctionService.deleteCookie('kmmsep_user');
-        $scope.signedIn = false;
-        $scope.userName = undefined;
+    $scope.signOut = function (ev) {
+        var confirm = $mdDialog.confirm()
+            .title('Ви впевнені, що хочете вийти?')
+            .targetEvent(ev)
+            .ok('Так, звісно')
+            .cancel('Ні');
+        $mdDialog.show(confirm)
+            .then(function () {
+                CommonFunctionService.deleteCookie('kmmsep_user');
+                $scope.signedIn = false;
+                $scope.userName = undefined;
+            })
     }
 
     $scope.$watch('userName', function (newValue, oldValue) {
@@ -114,8 +122,11 @@
         if (newValue != oldValue) {
             if (newValue != undefined) {
                 $scope.greetingMessage = createGreeting(newValue);
+                $scope.signedIn = true;
+                $scope.$apply;
             } else {
                 $scope.greetingMessage = 'Увійти';
+                $scope.signedIn = false;;
             }
         }
     })
